scripts/requirements-indexing.txt
langchain-text-splitters>=0.0.1
langchain-core>=0.1.43
langchain-openai>=0.1.2
langchain-pinecone>=0.1.1
pinecone-client>=3.0.0
